describe api, bll, dal for V3

+++ src/PriceCalculator.Api/Responses/V1/CargoResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace PriceCalculator.Api.Responses.V1;
 
 public record CargoResponse(
-    double Volume
+    decimal Volume
     );
